Fix Guava performance regression
Since 24.1, Guava included a change that made ImmutableSet.Builder with unspecified expected size allocate a lot more memory for many items. We haven't bumped into this problem before, because we used the Android variant of Guava that somehow didn't contain the regressing change. With moving to the Java 8 variant we now have to deal with this problem.